Finished implementing the bill validator reset.
Just need to figure out what the standard means when it gives the
bitvector definition for bill enabling; if I want to enable bills 15 and
3, do I send 8008H, 0880H, 0110H, or 1001H? All of these make sense,
depending on how you read the standard. My intuition says 8008, in which
case all that happens is a reverse before the shift-and-add, but 0110
can also make sense, in which case you reverse, split the list into two
pieces, and reverse each piece.